+++ ZF/ZFCore/zfsrc/ZFCore/ZFStyleLoad.cpp
@@ -86,22 +86,25 @@ ZFMETHOD_FUNC_DEFINE_3(zfbool, ZFStyleLoad,
 ZFMETHOD_FUNC_DEFINE_1(zfbool, ZFStyleLoad,
                        ZFMP_IN(const ZFSerializableData &, serializableData))
 {
-    ZFStyleChangeBegin();
-    for(zfindex i = 0; i < serializableData.elementCount(); ++i)
+    if(serializableData.elementCount() > 0)
     {
-        const ZFSerializableData &child = serializableData.elementAtIndex(i);
-        const zfchar *styleKey = ZFSerializableUtil::checkPropertyName(child);
-        if(styleKey != zfnull)
+        ZFStyleChangeBegin();
+        for(zfindex i = 0; i < serializableData.elementCount(); ++i)
         {
-            zfautoObject styleValueHolder = ZFObjectFromData(child);
-            ZFStyleable *styleValue = styleValueHolder;
-            if(styleValue != zfnull)
+            const ZFSerializableData &child = serializableData.elementAtIndex(i);
+            const zfchar *styleKey = ZFSerializableUtil::checkPropertyName(child);
+            if(styleKey != zfnull)
             {
-                ZFStyleSet(styleKey, styleValue);
+                zfautoObject styleValueHolder = ZFObjectFromData(child);
+                ZFStyleable *styleValue = styleValueHolder;
+                if(styleValue != zfnull)
+                {
+                    ZFStyleSet(styleKey, styleValue);
+                }
             }
         }
+        ZFStyleChangeEnd();
     }
-    ZFStyleChangeEnd();
     return zftrue;
 }