+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * optimize dynamic register impl
 * optimize lua implicit conversion
 * optimize lua path info impl
+* support output param for generic invoker
 * formalize api names
 
 # 0.3.0